Ray tracing demo: Combat tangent-space orientation "swap"
There was an issue when the tangent-space for randomDiskPoint
computation changed orientation/swapped in mid-screen the change in the
random distribution became noticeable on the final render.
We now use a simple solution taking the camera's "up" vector as the
tangent. This solution has issues, too, when the computed light
direction and the up vector align (=singularity), but we will stick with
it for now.