Ray tracing demo: Added hybrid rasterization + ray tracing demo
The idea of using hybrid rasterization + ray tracing approach is not
new. It allows to use the rasterizer for the first bounce of path
tracing to rasterize the (possibly complex) scene as usual with all
optimizations appricable to rasterization (viewing frustum culling,
occlusion culling, ...), store the view depth/position and normal in an
A-buffer. Later, that A-buffer can be sampled by the ray tracing shader
to compute shadow rays and reflected rays and continue with ray tracing
as usual.