Ray tracing demo: New "photon mapping" demo
This demo uses a cube map array texture to hold a "photon map" for each
of the boxes in the scene. Within a computer shader light rays are being
shot into the scene and whenever they hit a box the texel coordinate is
computed and the "photon" is stored in the corresponding face and layer
of the cube map array image.
Afterwards, the scene is rasterized and the cube map array is sampled
via a samplerCubeArray. The boxes are rendered via hardware instancing
and the layer of the cube map array is obtained via the gl_InstanceID.

All in all I'm kinda proud of this one, since it was written completely
from scratch without any reference.