Ray tracing demo: Query image binding point for image2D uniforms
The shader can specify, via the "binding" layout qualifier, the binding
point of an image uniform. This is the argument to glBindImageTexture to
bind an image of a texture to that point.
In order to query that, we use glGetUniform, since for image uniforms,
that "binding" value is the actual value of the uniform.