Ray tracing demo: Clear texture using pixel unpack
I wanted to clear the texture without using a host-allocated ByteBuffer.
OpenGL 4.4 has support for this with clearTexImage, but we want to stick
to 4.3 as that seems to be the limit for Kelper cards.
So, we are using a GPU-allocated buffer, clearing that with
clearBufferData and then uploading it to the texture via PBO and pixel
unpack.