Documentation: ClearBufferData and ClearBufferSubData
Make it clear that the "data" argument can be null, in which case the
storage of the buffer object is filled with zeros. I used only parts of
the OpenGL 4.3 Core specification text and changed it a bit to suit
LWJGL better (i.e. there are no "pointers" in Java and "buffer" could
mean both the buffer object storage as well as the Java Buffer instance
- it always means the buffer object storage in the spec).