Bindless textures demo: Use ARB_bindless_textures for many cube maps
This demo enhances the photon mapping demo by using not a CUBE_MAP_ARRAY
where each cube map has equal dimensions, but instead using
ARB_bindless_textures to create an "array" of single cube map textures
with dimensions proportional to the box it is applied to.
The addresses of each cube map are stored in a Uniform Buffer Object as
imageCube for image_load_store in the Compute Shader and as samplerCube
for sampling in the fragment shader.