Ray tracing demos: Added Demo20 using only OpenGL 2.0 features
It will use a RGB/XYZ 2D floating-point texture to hold the min/max corner coordinates of the boxes in a scene, instead of using an SSBO. We also need a simpler random function without relying on features of GLSL 3.30.
This all helps in porting the example to OpenGL ES 2.0 / WebGL 1.0 to be ultimately run on the web.