Add ray tracing demo with triangle meshes
On our way to a usable ray tracer this one now can trace real (albeit
small) triangle meshes. Currently no spatial acceleration structures
such as a kd-tree are being used.
A solution based on Stackless kd-tree traversal however is in the
making.

AtomicDemo: Set a random seed to the atomic counter buffer to avoid
patterns due to integer overflow when the number of pixels is a divisor
of MAX_INT.