Replace per-function address checks with a null check on the function provider instance. Fixes #54 Fixes #66
The per-function checks were pointless; we already validate function addresses when creating the provider instance and, if an unavailable extension was used, we'd always get an NPE when accessing the function address.

This change has two more benefits:

  a) we can throw a custom exception with a proper description when an unavailable extension is used
  b) we generate much less bytecode

Functions marked with DependsOn or ignoreMissing continue to validate the function address (in addition to the provider instance).