Replace optimistic capabilities with static capabilities (optionally enabled)
More testing showed that the optimistic implementation was not as efficient as the new static implementation and was slower than the thread-local implementation when an incompatible context was detected.

All capabilities now use the thread-local implementations by default. There are new Configuration options that enable the static capabilities. If Configuration.DEBUG is enabled, capabilities are checked when set and a warning will be printed if incompatible contexts are detected.