Major type system refactoring, make structs values instead of pointers
This fixes the biggest issue with the initial design. It made sense at the time because structs were always mapped to ByteBuffer (Java cannot support passing structs by value).

The new type system is sane. It's easier to understand and the code is simplified in several places. There are no changes to the generated code. Bonus fix: pointer to const pointer now works as expected.