perf: do not sort function pointers by name
This makes it more likely that for functions that are commonly used
together, their respective pointers will be closer in memory, avoiding
unnecessary cache misses.

Ideally, a log could be captured from a real application that would then
be used to optimally pack useful function pointers, in a custom LWJGL
build for that application.