perf(core): minimize inlining budget
This is a major refactoring in an attempt to minimize how many levels
LWJGL consumes out of Hotspot's default MaxInlineLevel=9. In a deep call
stack of application+LWJGL code, it may unlock additional opportunities
for optimizations. This is especially important when objects (e.g.
buffers) are passed between application and LWJGL, since crossing the
maximum inline threshold will cause EA to mark such objects as escaping
and Scalar Replacement won't be possible. For a practical example where
this trivially happens, see the previous memSlice version on Java 9 or
higher. Summary of changes:

- Implemented slice/duplicate of java.nio buffers using Unsafe.
- Implemented memSlice with offset without mutating the source buffer,
  making it thread-safe.
- Implemented struct and struct buffer creation using Unsafe.
- Implemented a number of workarounds to make struct buffer iteration
  garbage free on Hotspot. This applies to standard for-loop iteration,
  enhanced-for-loop and forEach. Streams are not garbage free on Hotspot
  but they are on Graal.
- Re-tuned string codecs for Java 8-11.
- Re-tuned memcpy/memset for Java 8-11.
- Manually inlined performance-sensitive code paths that were
  unnecessarily deep and the code reuse penalty is tiny.
- Unsafe is now required, there is no JNI fall-back.