feat(core): add support for native resources
These are files that can be bundled in the same way as shared libraries
in LWJGL natives JAR files.

The LibraryResource class can be used to resolve such resources. It uses
the same mechanism as Library, which means that the SharedLibraryLoader
automatically extracts resources to org.lwjgl.librarypath if necessary.