feat(generator): support C long type
This is required because on Windows a long is always 4 bytes, but on
other platforms it has the same size as the architecture's pointer size.

To maintain compatibility, the long type in Linux APIs remains mapped to
PrimitiveMapping.POINTER. The new long type currently shows up in the
Clang bindings only.

A new custom buffer (CLongBuffer) is introduced for pointers to long
values, but not yet added with this commit. Other than Linux APIs,
there's no pointer-to-long in any binding right now.