build: do not include dependency scopes in BOM
Discussed in #481.

This only takes care of removing dependency scopes from the BOM for the native artifacts (or more specifically, the workaround is updated to not add them). Adding another workaround to remove them from the "java" artifacts is possible but it's easier to update Gradle to a more recent version where this doesn't scopes are not published for dependency constraints.