fix(Vulkan): default instance/device API versions
The default instance & device API versions, when not specified by the
user, now match what the Vulkan specification says they should be.

- The default instance API version, when VkApplicationInfo is not
specified or when its apiVersion value is 0, is now VK_API_VERSION_1_0
instead of the maximum version supported by the instance.
- The default device API version, when not specified in the VkDevice
constructor, is now the minimum of the instance API version and the
API version supported by the physical device.
- The API version stored in VKCapabilitiesDevice is now the version
specified/derived in the VkDevice constructor, instead of always equal
to the instance API version.

Note that the highest API version the application targets, set with
VkApplicationInfo and stored in VKCapabilitiesInstance, can be higher
than the actual maximum version supported by the instance, as described
in the Vulkan specification.