fix APIUtil.apiGetManifestValue() for GraalVM native image build
The APIUtil.apiGetManifestValue() method doesn't work in a native image
generated by GraalVM, as ClassLoader.getResource() doesn't return a URL
starting with "jar:" as expected when running in standard JVM
environment. Instead, in native image which is run by an embedded
SubstrateVM, ClassLoader.getResource() returns a URL starting with
"resource:", for all resources in the classpath.

This fix would make APIUtil.apiGetManifestValue() work in both standard
JVM as well as in native image generated by GraalVM.