perf(core): simplify text encoding
The updated code shape works nicely across JVM versions and equally well
on both 32-bit and 64-bit JVMs.

It is also a workaround for an Unsafe-related C2 bug in Hotspot, that
causes silent data corruption on x86 with the server JVM. Reproducible
on JDK 8 up to 12, fixed on 13 and higher.