fix(Vulkan): always respect noautovalidity property. Close #702
Not happy with this fix, since it's obviously wrong in a few places
(e.g. the first three VkAllocationCallbacks function pointers).

LWJGL uses the "optional" property to mark parameters/members as
nullable, but it is missing in cases like the one reported with #702.
However, "noautovalidity" is consistent and LWJGL now interprets it
literally and does not perform any kind of validation, including
nullability. This may be reverted if the Vulkan-Docs situation changes
in the future.