feat(generator): automatic libffi downcalls
Functions that return or pass structs by value must be called with
either custom JNI code or libffi. When a module contains only a few such
functions and a JNI shared library would not otherwise be necessary for
it to work (the library is either provided by the system or it is a
non-JNI library built by LWJGL-CI), libffi is preferred. This avoids
having to bundle yet another shared library just for calling a few
functions.

This commit replaces custom code blocks (hand-written and error-prone)
with automatic libffi downcall generation which integrates cleanly with
other generator features.