perf(core) disable allocation stacktraces option
The DEBUG_MEMORY_ALLOCATOR_FAST option disables stacktrace generation
for tracked memory allocations, thus reducing DEBUG_MEMORY_ALLOCATOR's
overhead dramatically.

When enabled, memory leaks are detected and reported as usual, but
without an associated stacktrace. The low overhead enables constant
usage of the debug allocator during development. The option can be
disabled only when an actual memory leak is detected, to easily find the
offending memory allocation.

The option is also dynamic and can be used to activate stacktrace
generation in specific portions of the application, which may be suspect
for memory leaks.