feat(sdl) add SDL_main support
As an alternate way of structuring SDL apps.

The docs suggested it was not usable from other languages, but
https://github.com/libsdl-org/SDL/issues/14513 indicates a valid way of
using the main callbacks across an FFI boundary.

The callbacks don't do much on Windows/Linux/macOS. Mostly they handle
calling the iterate function while the window is resized on Windows. But
they might become more important in the future, and are the recommended
way of using SDL.